jobs/document-embedding/pyproject.toml
[project]
name = "document-embedding"
version = "0.1.0"
description = "Azure Container App Job for generating document embeddings"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.19.0",
    "azure-storage-queue>=12.0.0",
    "techpubs-core[embeddings]",
]

[tool.uv.sources]
techpubs-core = { workspace = true }
